+++ node/src/types/kernel/api/notebook/removeNotebook/index.d.ts
@@ -17,11 +17,5 @@
 
 
 /* directories */
-export * as closeNotebook from "./closeNotebook";
-export * as getNotebookConf from "./getNotebookConf";
-export * as lsNotebooks from "./lsNotebooks";
-export * as openNotebook from "./openNotebook";
-export * as removeNotebook from "./removeNotebook";
-export * as renameNotebook from "./renameNotebook";
 
 /* flles */