+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - 添加 API `/api/notebook/getNotebookConf` | Add API `/api/notebook/getNotebookConf`
 - 添加 API `/api/notebook/closeNotebook` | Add API `/api/notebook/closeNotebook`
 - 添加 API `/api/notebook/openNotebook` | Add API `/api/notebook/openNotebook`
+- 添加 API `/api/notebook/renameNotebook` | Add API `/api/notebook/renameNotebook`
 
 ## v0.0.4 / 2023-06-29